Update to ns-3.42 (#104)
* mmwave and lte changes

* remainder of src changes

* non-src folder changes

* update deprecated valgrind command options

* updated utils script

* fix leak and enum init in mmwave module

* updated build-support folder

* update test duration and type enums

* solve compiler warnings

* update python bindings and add docker doc

+++ src/network/model/packet-metadata.cc
@@ -493,7 +493,7 @@ PacketMetadata::ReplaceTail(PacketMetadata::SmallItem* item,
         AppendValue(extraItem->fragmentEnd, buffer);
         buffer += fragEndSize;
         Append32(extraItem->packetUid, buffer);
-        m_used = std::max(m_used, (uint16_t)(buffer - &m_data->m_data[0]));
+        m_used = std::max(m_used, static_cast<uint32_t>(buffer - &m_data->m_data[0]));
         m_data->m_dirtyEnd = m_used;
         return;
     }
@@ -1329,7 +1329,7 @@ PacketMetadata::AddToRawU8(const uint8_t& data, uint8_t* start, uint8_t* current
 {
     NS_LOG_FUNCTION(static_cast<uint32_t>(data) << &start << &current << maxSize);
     // First check buffer overflow
-    if (static_cast<uint32_t>((current + sizeof(uint8_t) - start)) > maxSize)
+    if (static_cast<uint32_t>(current + sizeof(uint8_t) - start) > maxSize)
     {
         return nullptr;
     }
@@ -1345,7 +1345,7 @@ PacketMetadata::AddToRawU16(const uint16_t& data,
 {
     NS_LOG_FUNCTION(data << &start << &current << maxSize);
     // First check buffer overflow
-    if (static_cast<uint32_t>((current + sizeof(uint16_t) - start)) > maxSize)
+    if (static_cast<uint32_t>(current + sizeof(uint16_t) - start) > maxSize)
     {
         return nullptr;
     }
@@ -1361,7 +1361,7 @@ PacketMetadata::AddToRawU32(const uint32_t& data,
 {
     NS_LOG_FUNCTION(data << &start << &current << maxSize);
     // First check buffer overflow
-    if (static_cast<uint32_t>((current + sizeof(uint32_t) - start)) > maxSize)
+    if (static_cast<uint32_t>(current + sizeof(uint32_t) - start) > maxSize)
     {
         return nullptr;
     }
@@ -1377,7 +1377,7 @@ PacketMetadata::AddToRawU64(const uint64_t& data,
 {
     NS_LOG_FUNCTION(data << &start << &current << maxSize);
     // First check buffer overflow
-    if (static_cast<uint32_t>((current + sizeof(uint64_t) - start)) > maxSize)
+    if (static_cast<uint32_t>(current + sizeof(uint64_t) - start) > maxSize)
     {
         return nullptr;
     }
@@ -1394,7 +1394,7 @@ PacketMetadata::AddToRaw(const uint8_t* data,
 {
     NS_LOG_FUNCTION(&data << dataSize << &start << &current << maxSize);
     // First check buffer overflow
-    if (static_cast<uint32_t>((current + dataSize - start)) > maxSize)
+    if (static_cast<uint32_t>(current + dataSize - start) > maxSize)
     {
         return nullptr;
     }
@@ -1410,7 +1410,7 @@ PacketMetadata::ReadFromRawU8(uint8_t& data,
 {
     NS_LOG_FUNCTION(static_cast<uint32_t>(data) << &start << &current << maxSize);
     // First check buffer underflow
-    if (static_cast<uint32_t>((current + sizeof(uint8_t) - start)) > maxSize)
+    if (static_cast<uint32_t>(current + sizeof(uint8_t) - start) > maxSize)
     {
         return nullptr;
     }
@@ -1426,7 +1426,7 @@ PacketMetadata::ReadFromRawU16(uint16_t& data,
 {
     NS_LOG_FUNCTION(data << &start << &current << maxSize);
     // First check buffer underflow
-    if (static_cast<uint32_t>((current + sizeof(uint16_t) - start)) > maxSize)
+    if (static_cast<uint32_t>(current + sizeof(uint16_t) - start) > maxSize)
     {
         return nullptr;
     }
@@ -1442,7 +1442,7 @@ PacketMetadata::ReadFromRawU32(uint32_t& data,
 {
     NS_LOG_FUNCTION(data << &start << &current << maxSize);
     // First check buffer underflow
-    if (static_cast<uint32_t>((current + sizeof(uint32_t) - start)) > maxSize)
+    if (static_cast<uint32_t>(current + sizeof(uint32_t) - start) > maxSize)
     {
         return nullptr;
     }
@@ -1458,7 +1458,7 @@ PacketMetadata::ReadFromRawU64(uint64_t& data,
 {
     NS_LOG_FUNCTION(data << &start << &current << maxSize);
     // First check buffer underflow
-    if ((uint32_t)((current + sizeof(uint64_t) - start)) > maxSize)
+    if ((uint32_t)(current + sizeof(uint64_t) - start) > maxSize)
     {
         return nullptr;
     }
